refactor: create logo component for navigation
