+++ src/components/Logo/Logo.tsx
@@ -1,13 +1,10 @@
-import React, {useContext} from 'react';
+import React from 'react';
 import Link from 'gatsby-link';
-import styled, {ThemeContext} from 'styled-components';
+import styled from 'styled-components';
 import {addRemToProperty} from '../../styles/shared';
+import Circle from './Circle';
 
-const RelativeWrapper = styled.div`
-  position: relative;
-`;
-
-const CircleBackground = styled.div`
+const Background = styled.div`
   position: absolute;
   top: 50%;
   left: 50%;
@@ -21,7 +18,7 @@ const CircleBackground = styled.div`
   will-change: transform;
 `;
 
-const CircleWrapper = styled.div`
+const Wrapper = styled.div`
   position: fixed;
   top: 0;
   left: 0;
@@ -54,42 +51,36 @@ const CircleWrapper = styled.div`
     height: auto;
   }
 
-  a:hover + ${CircleBackground} {
+  a:hover + ${Background} {
     opacity: 0;
     transform: translate(-50%, -50%) scale(7.5);
     transition: transform 1s cubic-bezier(0.45, 0, 0.1, 1), opacity 0.8s ease;
   }
 `;
-type CircleProps = {
-  readonly fill?: string;
-  readonly width?: number;
-  readonly height?: number;
-};
-const Circle: React.FC<CircleProps> = ({fill, width = 22, height = 22}) => (
-  <svg
-    width={width}
-    height={height}
-    viewBox="0 0 22 22"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <path
-      d="M11 21.5C5.201 21.5.5 16.799.5 11S5.201.5 11 .5 21.5 5.201 21.5 11 16.799 21.5 11 21.5zm0-1a9.5 9.5 0 1 0 0-19 9.5 9.5 0 0 0 0 19z"
-      fill={fill}
-    />
-  </svg>
-);
 
-const Logo = () => {
-  const themeContext = useContext(ThemeContext);
+const RelativeWrapper = styled.div`
+  position: relative;
+`;
+
+
+
+
+interface LogoProps {
+  fillColor: string;
+  link?: string;
+}
+
+const Logo: React.FC<LogoProps> = ({link = '/', fillColor = '#ffffff'}) => {
+
   return (
-    <CircleWrapper>
+    <Wrapper>
       <RelativeWrapper>
-        <Link to="/">
-          <Circle fill={themeContext.colors.lightest} />
+        <Link to={link}>
+          <Circle fill={fillColor} />
         </Link>
-        <CircleBackground />
+        <Background />
       </RelativeWrapper>
-    </CircleWrapper>
+    </Wrapper>
   );
 };
