+++ src/components/Logo/index.ts
@@ -1,0 +1,3 @@
+import LogoContainer from './LogoContainer';
+
+export default LogoContainer;
